requirements.txt
# Tested with Python 3.9.6

pandas==2.3.1
numpy==2.0.2
matplotlib==3.9.4
tabulate==0.9.0
fhir.resources==7.1.0
langchain==0.3.27
langchain-core==0.3.83
langchain-community==0.3.31
langchain-experimental==0.3.4
langchain-openai
openai==2.20.0
eval_type_backport==0.3.1
statsmodels==0.14.5